refactor: move invitation files to dedicated folder
