+++ app/Http/Controllers/EventController.php
@@ -6,33 +6,14 @@ use App\Enums\InvitationStatus;
 use App\Models\Event;
 use App\Models\Invitation;
 use Barryvdh\DomPDF\Facade\Pdf;
-use SimpleSoftwareIO\QrCode\Facades\QrCode;
 
 class EventController
 {
-    public function index(Invitation $invitation)
-    {
-        $event = $invitation->event;
-        $template = 'templates.'.$event->template->view;
-
-        return view('event.index', compact('event', 'invitation', 'template'));
-    }
-
-    public function downloadTickets(Event $event, Invitation $invitation)
-    {
-        $qr = base64_encode(QrCode::size(200)->generate($invitation->code));
-
-        $pdf = Pdf::loadView('event.tickets', compact('event', 'invitation', 'qr'))
-            ->setPaper([0, 0, 375, 590]);
-
-        return $pdf->stream('tickets.pdf');
-    }
-
     public function preview(Event $event)
     {
         $template = 'templates.'.$event->template->view;
 
-        return view('event.index', compact('event', 'template'));
+        return view('invitation.index', compact('event', 'template'));
     }
 
     public function downloadInvitationsList(Event $event)
@@ -46,7 +27,7 @@ class EventController
             ->get()
             ->groupBy('table');
 
-        $pdf = Pdf::loadView('event.invitations', compact('event', 'tables'))
+        $pdf = Pdf::loadView('event.invitations-list', compact('event', 'tables'))
             ->setOption(['fontDir' => sys_get_temp_dir(), 'isPhpEnabled' => true])
             ->setPaper('Letter');
