--- resources/views/checkout-success.blade.php
@@ -1,5 +1,0 @@
-<div>
-    <pre>
-    {{ $session }}
-    </pre>
-</div>
