+++ resources/views/event/invitations-list.blade.php
