+++ resources/views/invitation/index.blade.php
