+++ resources/views/livewire/event/auth-event.blade.php
