+++ resources/views/livewire/invitation/show-qr-code.blade.php
